Rename project_details_page back to create_project_details to merge from develop.
